\name{notdocumentedyet}
\alias{notdocumentedyet}

\alias{dbexist}
\alias{TableExist}
\alias{DeleteObjFromDb}
\alias{GetFileFromDb}
\alias{GetObjFromDb}
\alias{GetScriptFromDb}
\alias{ObjectExists}
\alias{ObjToStr}
\alias{StrToObj}
\alias{SplitStr}
\alias{StoreFileToDb}
\alias{StoreObjToDb}
\alias{StoreScriptToDb}
\alias{UnSplitStr}

%
%
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Undocumented and Internally Used Functions and Classes }
\description{
Currently undocumented functions

}
%\usage{
%normal1(lmean="identity", lsd="loge", zero=NULL)
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%  \item{lmean}{
%  Link function applied to the mean.
%  See \code{\link{Links}} for more choices.
%
%  }
%}
\details{
  In the \pkg{pgobjects} package there are currently many
  objects/methods/classes which are currently internal and/or
  undocumented.

}
\value{
  Each objects/methods/classes may or may not have its own individual value.
  These will be documented over time.

}
%\references{
%}

\author{ Vectrun }
%\note{
%
%}
%\seealso{
%    \code{gaussianff},
%    \code{\link{posnormal1}}.
%}
%\examples{
%}
\keyword{undocumented}







