\name{pgobjects-package}
\alias{pgobjects-package}
\alias{pgobjects}
\docType{package}
\title{Postgresql - R objects
}
\description{
    Package to maintain R objects within a postgresql database
}
\details{
\tabular{ll}{
Package: \tab pgobjects\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-03-15\cr
License: \tab  GPL v3.0\cr
LazyLoad: \tab yes\cr
}
With this package it is possible to store R objects in a postgresql
database. These objects can then be exchanged between (remote) R
sessions. Since everything is a object in R it is possible to store
data, functions, classes, and even files in the database
}
\author{
Vectrun

Maintainer: Vectrun <vectrun@gmail.com>
}
\references{
    none
}
\keyword{ package }
\seealso{
}
\examples{
}
