% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgObjectsClose.R
\name{PgObjectsClose}
\alias{PgObjectsClose}
\title{Close database connectio object}
\usage{
PgObjectsClose(handle = get(getOption("pgobject.dbhandle")))
}
\arguments{
\item{handle}{dbhandle of open connection

The dbhandle is retrieved from the R options, usually you don't
supply this parameter.}
}
\description{
This function closes the database connection opened by
\code{\link{PgObjectsInit}}
}
