% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgObjectsInit.R
\name{PgObjectsInit}
\alias{PgObjectsInit}
\title{Initialize pgobjects database connection}
\usage{
PgObjectsInit(dbname, user = NA, host = "localhost", passwd = "",
  dbhandle = "dbh", port = 5432, schema = "public", chunksize = 7168)
}
\arguments{
\item{dbname}{Name of the database to connect to}

\item{user}{Username}

\item{host}{hostname, default 'localhost'}

\item{passwd}{Password for the database (if any)}

\item{dbhandle}{Name of the dbhandle in .Globalenv}

\item{port}{Portnumber of database, defaults to 5432}

\item{schema}{Which database schema to use, defaults to 'public'}

\item{chunksize}{Chunksize, size of individual chunks}
}
\description{
Function to open connection to a database and returns a database handle
}
\details{
This function opens a database connection using the postgresql
driver. It returns a database handle which is used within other
functions of the pgobjects package.
}
