% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql}
\alias{sql}
\title{SQL wraper function}
\usage{
sql(query, verbose = FALSE, errors = TRUE, dbhandle = NA)
}
\arguments{
\item{query}{The SQL query}

\item{verbose}{If true give verbose output}

\item{errors}{If True give errors as warning}
}
\description{
Wrapper function to execute sql queries on a postgres database
}
\details{
this function is the workhorse of the pgobjects library, it runs
sql queries at the RDBMS and returns the result. This function
tries to catch any error and then returns NA. It is up to the user 
to test te result of the function.
}
\seealso{
\code{\link{PgObjectsInit}}
}
